--- Assets/Prefabs/BarreDeVie.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 084fc32681e8b9d47b79816c58e89d69
-timeCreated: 1529227866
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
