--- Assets/Prefabs/Effects/Fireball.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 2092e1c873754b84195f228b1012a352
-timeCreated: 1523559234
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
