--- Assets/Scripts/HeroesScript/BarreDeVie.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: e52f2b41e4b35664cbe6e32c9d367fb6
-timeCreated: 1529224815
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
