+++ Assets/Scripts/HeroesScript/Init.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 760563ab709ae7945a6117ae51d7e23a
-timeCreated: 1531416815
+guid: 28fe0d4beb7d9be4fa0ef9364e571e02
+timeCreated: 1519161054
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
